Create Code.gs
Currently this script cannot find the specified folder "Bloggy" anymore as Google altered something a few months back.  When I get a chance, I will update the "DocsList" code so it can properly find the folder.  When that is done, I am going to flip on HTML publishing so that html can be entered from the user interface of the app.